move bindDepsToFn to helpers module
